requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.2
pypdf==4.3.1
python-docx==1.1.2
openai==0.28.1
plotly
pandas